api/pyproject.toml
[tool.poetry]
name = "landmarks-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.5"


[tool.poetry.group.api.dependencies]
absl-py = "1.3.0"
aiohttp = "3.8.3"
aiohttp_cors = "0.7.0"
aiosignal = "1.2.0"
antlr4-python3-runtime = "4.8"
anyio = "3.6.2"
argon2-cffi = "21.3.0"
argon2-cffi-bindings = "21.2.0"
asttokens = "2.1.0"
async-timeout = "4.0.2"
attrs = "22.1.0"
ax-platform = "0.2.8"
backcall = "0.2.0"
beautifulsoup4 = "4.11.1"
bleach = "5.0.1"
blessed = "1.19.1"
boto3 = "1.15.6"
botocore = "1.18.18"
botorch = "0.7.2"
cachetools = "5.2.0"
certifi = "2022.9.24"
cffi = "1.15.1"
charset-normalizer = "2.1.1"
click = "8.0.4"
cloudpickle = "2.2.0"
colorful = "0.5.4"
colorlog = "6.7.0"
contourpy = "1.0.6"
cycler = "0.11.0"
debugpy = "1.6.3"
decorator = "5.1.1"
defusedxml = "0.7.1"
distlib = "0.3.6"
entrypoints = "0.4"
executing = "1.2.0"
faiss-gpu = "1.6.5"
fastjsonschema = "2.16.2"
filelock = "3.8.0"
fonttools = "4.38.0"
frozenlist = "1.3.1"
google-api-core = "2.10.2"
google-auth = "2.14.1"
google-auth-oauthlib = "0.4.6"
googleapis-common-protos = "1.56.4"
gpustat = "1.0.0"
gpytorch = "1.9.0"
grpcio = "1.50.0"
huggingface-hub = "0.10.1"
hydra-ax-sweeper = "1.1.0"
hydra-colorlog = "1.0.1"
hydra-core = "1.0.6"
hydra-ray-launcher = "0.1.2"
idna = "3.4"
importlib-metadata = "5.0.0"
ipykernel = "6.17.0"
ipython = "8.6.0"
ipython_genutils = "0.2.0"
ipywidgets = "8.0.2"
jedi = "0.18.1"
Jinja2 = "3.1.2"
jmespath = "0.10.0"
joblib = "1.2.0"
jsonschema = "4.17.0"
jupyter = "1.0.0"
jupyter-console = "6.4.4"
jupyter-server = "1.23.0"
jupyter-client = "7.4.4"
jupyter-core = "4.11.2"
jupyterlab-pygments = "0.2.2"
jupyterlab-widgets = "3.0.3"
kiwisolver = "1.4.4"
linear-operator = "0.1.1"
Markdown = "3.4.1"
MarkupSafe = "2.1.1"
matplotlib = "3.6.2"
matplotlib-inline = "0.1.6"
mistune = "2.0.4"
msgpack = "1.0.4"
multidict = "6.0.2"
multipledispatch = "0.6.0"
nbclassic = "0.4.8"
nbclient = "0.7.0"
nbconvert = "7.2.3"
nbformat = "5.7.0"
nest-asyncio = "1.5.6"
notebook = "6.5.2"
notebook-shim = "0.2.2"
numpy = "1.23.4"
nvidia-cublas-cu11 = "11.10.3.66"
nvidia-cuda-nvrtc-cu11 = "11.7.99"
nvidia-cuda-runtime-cu11 = "11.7.99"
nvidia-cudnn-cu11 = "8.5.0.96"
nvidia-ml-py = "11.495.46"
oauthlib = "3.2.2"
omegaconf = "2.0.6"
opencensus = "0.11.0"
opencensus-context = "0.1.3"
opencv-python = "4.6.0.66"
opt-einsum = "3.3.0"
packaging = "21.3"
pandas = "1.5.1"
pandocfilters = "1.5.0"
parso = "0.8.3"
pexpect = "4.8.0"
pickle5 = "0.0.11"
pickleshare = "0.7.5"
Pillow = "9.3.0"
platformdirs = "2.5.3"
plotly = "5.11.0"
prometheus-client = "0.13.1"
prompt-toolkit = "3.0.32"
protobuf = "3.19.6"
psutil = "5.9.4"
ptyprocess = "0.7.0"
pure-eval = "0.2.2"
py-spy = "0.3.14"
pyasn1 = "0.4.8"
pyasn1-modules = "0.2.8"
pycparser = "2.21"
pydantic = "1.10.2"
Pygments = "2.13.0"
pyparsing = "3.0.9"
pyro-api = "0.1.2"
pyro-ppl = "1.8.2"
pyrsistent = "0.19.2"
python-dateutil = "2.8.2"
pytorch-metric-learning = "1.6.3"
pytz = "2022.6"
PyYAML = "6.0"
pyzmq = "24.0.1"
qtconsole = "5.4.0"
QtPy = "2.3.0"
ray = "2.1.0"
requests = "2.28.1"
requests-oauthlib = "1.3.1"
rsa = "4.9"
s3transfer = "0.3.7"
scikit-learn = "1.1.3"
scipy = "1.9.3"
Send2Trash = "1.8.0"
six = "1.16.0"
smart-open = "6.2.0"
sniffio = "1.3.0"
soupsieve = "2.3.2.post1"
stack-data = "0.6.0"
tabulate = "0.9.0"
tenacity = "8.1.0"
tensorboard = "2.10.1"
tensorboard-data-server = "0.6.1"
tensorboard-plugin-wit = "1.8.1"
tensorboardX = "2.5.1"
terminado = "0.17.0"
threadpoolctl = "3.1.0"
timm = "0.6.11"
tinycss2 = "1.2.1"
torch = "1.12.1"
torchvision = "0.13.1"
tornado = "6.2"
tqdm = "4.64.1"
traitlets = "5.5.0"
typeguard = "2.13.3"
typing-extensions = "4.4.0"
urllib3 = "1.25.11"
virtualenv = "20.16.6"
wcwidth = "0.2.5"
webencodings = "0.5.1"
websocket-client = "1.4.2"
Werkzeug = "2.2.2"
widgetsnbextension = "4.0.3"
yarl = "1.8.1"
zipp = "3.10.0"


[tool.poetry.group.web.dependencies]
streamlit = "^1.15.1"
Flask = "^2.2.2"
gunicorn = "^20.1.0"
jsonpickle = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
